pyproject.toml
[tool.poetry]
name = "todo"
version = "0.1.0"
description = ""
authors = ["Praveen Ilangovan <praveen.ilangovan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.4"}
beanie = "^1.27.0"
python-dotenv = "^1.0.1"

[tool.poetry.scripts]
start = "todo.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
